Move the raw analysis functions to a new file
